pyproject.toml
[project]
name = "ic"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicholas",email = "nicholasnas.nn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "ic", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
